+++ RELEASE-NOTES.md
@@ -39,6 +39,7 @@
 * Added image repository model in snowflake.yml.
 * Added `snow spcs service deploy` command.
 * Added notebooks to `snow object` commands.
+* Added `snow notebook deploy` command that allows creating a notebook using local file.
 
 ## Fixes and improvements